+++ packages/devtools-solana/jest.config.js
@@ -3,6 +3,7 @@ module.exports = {
     cache: false,
     reporters: [['github-actions', { silent: false }], 'default'],
     testEnvironment: 'node',
+    testTimeout: 15000,
     setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
     moduleNameMapper: {
         '^@/(.*)$': '<rootDir>/src/$1',