+++ packages/devtools-move/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @layerzerolabs/devtools-move
 
+## 1.0.15
+
+### Patch Changes
+
+- 06dc63b: quality of life improvements
+
 ## 1.0.14
 
 ### Patch Changes